Fixed override of font family for bold and italic formatting
By calling setFontFamilies with an empty list, it seems we're setting a
flag on the format that the font family has been explicitly set.

This made bold and italic formatting no longer respect the family of the
existing text, or in the case of Notes, a custom font configured by the
user.